refactor: CreateAhorcado
