+++ src/activity/components/createAhorcado/CreateAhorcado.module.css
@@ -54,423 +54,6 @@
   margin-bottom: 2rem;
 }
 
-/* Card Styles */
-.configCard,
-.difficultyCard,
-.attemptsCard,
-.previewCard {
-  background: white;
-  border-radius: 12px;
-  padding: 1.5rem;
-  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
-  border: 1px solid #e5e7eb;
-  transition: all 0.3s ease;
-}
-
-.configCard:hover,
-.difficultyCard:hover,
-.attemptsCard:hover,
-.previewCard:hover {
-  box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1);
-  transform: translateY(-2px);
-}
-
-.cardHeader {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-  margin-bottom: 1.5rem;
-  padding-bottom: 1rem;
-  border-bottom: 2px solid #f3f4f6;
-}
-
-.cardIcon {
-  font-size: 1.5rem;
-  color: #667eea;
-}
-
-.cardTitle {
-  font-size: 1.25rem;
-  font-weight: 600;
-  color: #1f2937;
-  margin: 0;
-}
-
-/* Form Styles */
-.formGroup {
-  margin-bottom: 1.5rem;
-}
-
-.label {
-  display: block;
-  font-weight: 600;
-  color: #374151;
-  margin-bottom: 0.5rem;
-  font-size: 0.875rem;
-}
-
-.required {
-  color: #ef4444;
-  margin-left: 0.25rem;
-}
-
-.wordInput {
-  font-family: "Courier New", monospace;
-  font-size: 1.1rem;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.1em;
-}
-
-.numberInput {
-  max-width: 120px;
-}
-
-.inputHelper {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: 0.5rem;
-  font-size: 0.75rem;
-  color: #6b7280;
-}
-
-.charCount {
-  color: #9ca3af;
-}
-
-.wordLength {
-  color: #059669;
-  font-weight: 500;
-}
-
-/* Suggestions */
-.suggestions {
-  margin-top: 1rem;
-  padding-top: 1rem;
-  border-top: 1px solid #e5e7eb;
-}
-
-.suggestionsLabel {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  font-size: 0.875rem;
-  font-weight: 600;
-  color: #374151;
-  margin-bottom: 0.75rem;
-}
-
-.suggestionIcon {
-  color: #f59e0b;
-}
-
-.suggestionsList {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-}
-
-.suggestionButton {
-  background: #f3f4f6;
-  border: 1px solid #d1d5db;
-  border-radius: 6px;
-  padding: 0.375rem 0.75rem;
-  font-size: 0.75rem;
-  font-weight: 500;
-  color: #374151;
-  cursor: pointer;
-  transition: all 0.2s ease;
-}
-
-.suggestionButton:hover {
-  background: #667eea;
-  color: white;
-  border-color: #667eea;
-  transform: translateY(-1px);
-}
-
-/* Difficulty Styles */
-.difficultyGrid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-  gap: 1rem;
-  margin-bottom: 1.5rem;
-}
-
-.difficultyButton {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-  padding: 1rem;
-  background: white;
-  border: 2px solid #e5e7eb;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.difficultyButton:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.difficultyButton.active {
-  background: #f8fafc;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.difficultyButton.active.hard {
-  border-color: #ef4444;
-  background: #fef2f2;
-}
-
-.difficultyButton.active.medium {
-  border-color: #f59e0b;
-  background: #fffbeb;
-}
-
-.difficultyButton.active.easy {
-  border-color: #10b981;
-  background: #f0fdf4;
-}
-
-.difficultyColor {
-  width: 16px;
-  height: 16px;
-  border-radius: 50%;
-}
-
-.difficultyColor.hard {
-  background: #ef4444;
-}
-
-.difficultyColor.medium {
-  background: #f59e0b;
-}
-
-.difficultyColor.easy {
-  background: #10b981;
-}
-
-.difficultyInfo {
-  display: flex;
-  flex-direction: column;
-}
-
-.difficultyLabel {
-  font-weight: 600;
-  color: #1f2937;
-}
-
-.difficultyErrors {
-  font-size: 0.75rem;
-  color: #6b7280;
-}
-
-.customDifficulty {
-  padding-top: 1rem;
-  border-top: 1px solid #e5e7eb;
-}
-
-/* Attempts Styles */
-.attemptsInfo {
-  color: #059669;
-  font-weight: 500;
-}
-
-.attemptsButtons {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.75rem;
-  margin-top: 1rem;
-}
-
-.attemptButton {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.75rem 1rem;
-  background: white;
-  border: 2px solid #e5e7eb;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  font-size: 0.875rem;
-  font-weight: 500;
-  color: #374151;
-}
-
-.attemptButton:hover {
-  border-color: #667eea;
-  color: #667eea;
-  transform: translateY(-1px);
-}
-
-.attemptButton.active {
-  background: #667eea;
-  border-color: #667eea;
-  color: white;
-}
-
-.attemptIcon {
-  font-size: 0.75rem;
-}
-
-/* Preview Styles */
-.previewBadge {
-  margin-left: auto;
-}
-
-.gamePreview {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 2rem;
-  align-items: start;
-}
-
-.hangmanDisplay {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: #f9fafb;
-  border-radius: 8px;
-  padding: 2rem;
-  min-height: 300px;
-}
-
-.hangmanArt {
-  font-family: "Courier New", monospace;
-  font-size: 1rem;
-  line-height: 1.2;
-  color: #374151;
-  margin: 0;
-  white-space: pre;
-  text-align: center;
-}
-
-.gameInfo {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-.wordDisplay {
-  text-align: center;
-}
-
-.wordTitle {
-  font-size: 1rem;
-  font-weight: 600;
-  color: #374151;
-  margin: 0 0 0.75rem 0;
-}
-
-.wordLetters {
-  font-family: "Courier New", monospace;
-  font-size: 2rem;
-  font-weight: 700;
-  color: #1f2937;
-  letter-spacing: 0.5rem;
-  background: #f3f4f6;
-  padding: 1rem;
-  border-radius: 8px;
-  border: 2px dashed #d1d5db;
-  min-height: 60px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.gameStats {
-  display: grid;
-  gap: 0.75rem;
-}
-
-.stat {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.75rem;
-  background: #f9fafb;
-  border-radius: 6px;
-}
-
-.statLabel {
-  font-weight: 500;
-  color: #6b7280;
-  font-size: 0.875rem;
-}
-
-.statValue {
-  font-weight: 700;
-  color: #1f2937;
-  font-size: 0.875rem;
-}
-
-.previewControls {
-  display: flex;
-  gap: 0.75rem;
-  align-items: flex-end;
-}
-
-.guessInput {
-  flex: 1;
-  max-width: 80px;
-  text-align: center;
-  font-family: "Courier New", monospace;
-  font-size: 1.25rem;
-  font-weight: 700;
-  text-transform: uppercase;
-}
-
-.guessButton {
-  white-space: nowrap;
-}
-
-.guessedLetters {
-  padding-top: 1rem;
-  border-top: 1px solid #e5e7eb;
-}
-
-.guessedLabel {
-  display: block;
-  font-size: 0.875rem;
-  font-weight: 600;
-  color: #374151;
-  margin-bottom: 0.5rem;
-}
-
-.lettersList {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-}
-
-.guessedLetter {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 32px;
-  height: 32px;
-  border-radius: 6px;
-  font-family: "Courier New", monospace;
-  font-weight: 700;
-  font-size: 0.875rem;
-}
-
-.guessedLetter.correct {
-  background: #dcfce7;
-  color: #166534;
-  border: 1px solid #bbf7d0;
-}
-
-.guessedLetter.incorrect {
-  background: #fee2e2;
-  color: #dc2626;
-  border: 1px solid #fecaca;
-}
-
 /* Footer */
 .footer {
   border-top: 1px solid #e5e7eb;
@@ -598,11 +181,6 @@
     padding: 0.5rem;
   }
 
-  .cardHeader {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 0.5rem;
-  }
 }
 
 /* Animation Classes */
@@ -653,7 +231,7 @@
 }
 
 /* Hover Effects */
-.configCard:hover .cardIcon,
+
 .difficultyCard:hover .cardIcon,
 .attemptsCard:hover .cardIcon,
 .previewCard:hover .cardIcon {
