+++ src/activity/components/createAhorcado/components/ahorcadoDifficultySelector/AhorcadoDificultySelector.module.css
@@ -1,0 +1,121 @@
+.difficultyCard{
+    background: white;
+    border-radius: 12px;
+    padding: 1.5rem;
+    box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
+    border: 1px solid #e5e7eb;
+    transition: all 0.3s ease;
+}
+.difficultyCard:hover{
+    box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1);
+    transform: translateY(-2px);
+}
+.cardHeader {
+    display: flex;
+    align-items: center;
+    gap: 0.75rem;
+    margin-bottom: 1.5rem;
+    padding-bottom: 1rem;
+    border-bottom: 2px solid #f3f4f6;
+}
+
+.cardIcon {
+  font-size: 1.5rem;
+  color: #667eea;
+}
+
+.cardTitle {
+  font-size: 1.25rem;
+  font-weight: 600;
+  color: #1f2937;
+  margin: 0;
+}
+
+.difficultyGrid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+  gap: 1rem;
+  margin-bottom: 1.5rem;
+}
+
+.difficultyButton {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+  padding: 1rem;
+  background: white;
+  border: 2px solid #e5e7eb;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: all 0.3s ease;
+}
+
+.difficultyButton:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.difficultyButton.active {
+  background: #f8fafc;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.difficultyButton.active.hard {
+  border-color: #ef4444;
+  background: #fef2f2;
+}
+
+.difficultyButton.active.medium {
+  border-color: #f59e0b;
+  background: #fffbeb;
+}
+
+.difficultyButton.active.easy {
+  border-color: #10b981;
+  background: #f0fdf4;
+}
+
+.difficultyColor {
+  width: 16px;
+  height: 16px;
+  border-radius: 50%;
+}
+
+.difficultyColor.hard {
+  background: #ef4444;
+}
+
+.difficultyColor.medium {
+  background: #f59e0b;
+}
+
+.difficultyColor.easy {
+  background: #10b981;
+}
+
+.difficultyInfo {
+  display: flex;
+  flex-direction: column;
+}
+
+.difficultyLabel {
+  font-weight: 600;
+  color: #1f2937;
+}
+
+.difficultyErrors {
+  font-size: 0.75rem;
+  color: #6b7280;
+}
+
+.customDifficulty {
+  padding-top: 1rem;
+  border-top: 1px solid #e5e7eb;
+}
+.label {
+  display: block;
+  font-weight: 600;
+  color: #374151;
+  margin-bottom: 0.5rem;
+  font-size: 0.875rem;
+}
