+++ src/activity/views/activityView/ActivityView.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import type React from "react";
 import { motion } from "framer-motion";
 import styles from "./ActivityView.module.css";
